requirements.txt
flask
tensorflow
opencv-python-headless
numpy
streamlit
requests
pillow
